importing all OFF category names
Before, only the English (or if not available, the first) category name was imported. Now, all names in all languages and all synonymous names in one language are imported. This required a change of the target database table structure, namely an additional category_names table. Also, some schema refactoring was done alongside.